+++ NEWS.md
@@ -1,5 +1,6 @@
 # 0.3.0 [not on CRAN yet]
 
+- improved error checking
 - ordering of groups in plot is now the same as the order of "groups" column in data
 - changed `colors` argument default to "color" (i.e. if a column `color` is present in your data, it will be used for coloring the events)
 - bugfix if data contains only one event
@@ -8,6 +9,8 @@
     + `title` (gives a title for the timeline)
     + `tooltips` (column name of data that contains individual tooltips)
     + `fontcolors` (column name of data that contains color of the event font)
+- bugfixes of levels of ranges (_-_-)
+- bugfix where events where not correctly categorized into their respective groups
 
 # 0.2.0
 - improved error checking